modules
^ models > modules/models
^ move created events to modules/libs/events.ts
^ move created categories to modules/libs/categories.ts
