+++ src/actions/commands/help/help.ts
@@ -1,7 +1,7 @@
 import { ChatInputCommandInteraction, SlashCommandBuilder } from 'discord.js'
 
-import { Action } from '../../../models/action'
-import { ActionEvents } from '../../../models/event'
+import { Action } from '../../../modules/models/action'
+import { ActionEvents } from '../../../modules/libs/events'
 
 import { Pages } from './models/pages.i'
