+++ src/actions/commands/status.ts
@@ -1,8 +1,8 @@
 import { ChatInputCommandInteraction, EmbedBuilder, SlashCommandBuilder } from 'discord.js'
 
-import { Action } from '../../models/action'
-import { ActionEvents } from '../../models/event'
-import { ActionCategories } from '../../models/category'
+import { Action } from '../../modules/models/action'
+import { ActionEvents } from '../../modules/libs/events'
+import { ActionCategories } from '../../modules/libs/categories'
 
 export default new Action({
   data: new SlashCommandBuilder()
