+++ src/actions/events/warn.ts
@@ -1,5 +1,5 @@
-import { Action } from '../../models/action'
-import { ActionEvents } from '../../models/event'
+import { Action } from '../../modules/models/action'
+import { ActionEvents } from '../../modules/libs/events'
 
 import logger from '../../utils/helpers/logger'
