+++ src/actions/test/commands/test-button.ts
@@ -6,8 +6,8 @@ import {
   SlashCommandBuilder,
 } from 'discord.js'
 
-import { Action } from '../../../models/action'
-import { ActionEvents } from '../../../models/event'
+import { Action } from '../../../modules/models/action'
+import { ActionEvents } from '../../../modules/libs/events'
 
 import logger from '../../../utils/helpers/logger'
