+++ src/actions/test/components/menu.ts
@@ -1,7 +1,7 @@
 import { StringSelectMenuInteraction } from 'discord.js'
 
-import { Action } from '../../../models/action'
-import { ActionEvents } from '../../../models/event'
+import { Action } from '../../../modules/models/action'
+import { ActionEvents } from '../../../modules/libs/events'
 
 import logger from '../../../utils/helpers/logger'
