+++ src/actions/test/messages/delete.ts
@@ -1,7 +1,7 @@
 import { Message } from 'discord.js'
 
-import { Action } from '../../../models/action'
-import { ActionEvents } from '../../../models/event'
+import { Action } from '../../../modules/models/action'
+import { ActionEvents } from '../../../modules/libs/events'
 
 import logger from '../../../utils/helpers/logger'
 
@@ -11,8 +11,8 @@ export default new Action({
   event: ActionEvents.MessageDelete,
 
   async execute(message: Message) {
-    const {author} = message
-    
+    const { author } = message
+
     logger.log(`${author.tag} delete a message`)
   },
 })
