+++ src/modules/models/category.ts
@@ -1,0 +1,5 @@
+export interface ICategory {
+  name: string
+  description?: string
+  emoji?: string
+}
