+++ src/modules/models/event.ts
@@ -1,0 +1,6 @@
+import { ClientEvents } from 'discord.js'
+
+export interface IEvent {
+  name: keyof ClientEvents
+  init: (...args: any) => any
+}
