+++ src/plugins/temporaryVoice/create.ts
@@ -7,7 +7,7 @@ import {
   VoiceState,
 } from 'discord.js'
 
-import { Action } from '../../models/action'
+import { Action } from '../../modules/models/action'
 
 import { validateAction } from '../../utils/helpers/validateAction'
