+++ src/plugins/ticketTool/components/delete.ts
@@ -1,6 +1,6 @@
 import { ButtonInteraction, Events } from 'discord.js'
 
-import { Action } from '../../../models/action'
+import { Action } from '../../../modules/models/action'
 
 import { validateAction } from '../../../utils/helpers/validateAction'
 import { deleteTicket } from '../utils/ticket/delete.i'
@@ -15,11 +15,7 @@ export default new Action({
   async init(interaction: ButtonInteraction) {
     if (this.data.name !== interaction.customId) return
 
-    const invalidation = validateAction(
-      this,
-      interaction.guild,
-      interaction.user
-    )
+    const invalidation = validateAction(this, interaction.guild, interaction.user)
 
     if (invalidation) {
       return await interaction.reply({
