+++ src/plugins/ticketTool/components/open.ts
@@ -1,6 +1,6 @@
 import { ButtonInteraction, Events } from 'discord.js'
 
-import { Action } from '../../../models/action'
+import { Action } from '../../../modules/models/action'
 
 import { validateAction } from '../../../utils/helpers/validateAction'
 import { openTicket } from '../utils/ticket/open.i'
@@ -15,11 +15,7 @@ export default new Action({
   async init(interaction: ButtonInteraction) {
     if (this.data.name !== interaction.customId) return
 
-    const invalidation = validateAction(
-      this,
-      interaction.guild,
-      interaction.user
-    )
+    const invalidation = validateAction(this, interaction.guild, interaction.user)
 
     if (invalidation) {
       return await interaction.reply({
