+++ src/utils/helpers/getCategoryName.ts
@@ -1,0 +1,8 @@
+import { ICategory } from '../../modules/models/category'
+
+export function getCategoryName(category: ICategory): string {
+  const name = []
+  if (category.emoji) name.push(category.emoji)
+  name.push(category.name)
+  return name.join('・')
+}
