+++ src/utils/helpers/getCommandUsage.ts
@@ -10,16 +10,14 @@ import { SlashCommandBuilder, SlashCommandSubcommandBuilder } from 'discord.js'
  * @param {SlashCommandBuilder} command - instance of SlashCommandBuilder
  * @returns {string[]} - returns array [command name, ...options] in right format
  */
-export function generateCommandUsage(command: SlashCommandBuilder): string[] {
+export function getCommandUsage(command: SlashCommandBuilder): string[] {
   const { name, options } = command
 
   const nameStr = `/${name}`
   const optionsStr = options.map((option) => {
     const { name, required } = option.toJSON()
 
-    return required || option instanceof SlashCommandSubcommandBuilder
-      ? `<${name}>`
-      : `(${name})`
+    return required || option instanceof SlashCommandSubcommandBuilder ? `<${name}>` : `(${name})`
   })
 
   return [nameStr, ...optionsStr]
