+++ src/utils/setup/loadPlugins.ts
@@ -1,8 +1,8 @@
 import path from 'path'
 import fs from 'fs'
 
-import { SkynetClient } from '../../models/client'
-import { Action } from '../../models/action'
+import { SkynetClient } from '../../modules/models/client'
+import { Action } from '../../modules/models/action'
 
 import { getFiles } from '../helpers/fileSystem'
 import logger from '../helpers/logger'
